ref(ui): Remove LightWeightNoProjectMessage
