+++ static/app/components/noProjectMessage.tsx
@@ -11,15 +11,16 @@ import {t} from 'app/locale';
 import ConfigStore from 'app/stores/configStore';
 import space from 'app/styles/space';
 import {Organization, Project} from 'app/types';
+import withProjects from 'app/utils/withProjects';
 
 type Props = React.PropsWithChildren<{
   organization: Organization;
-  projects?: Project[];
-  loadingProjects?: boolean;
+  projects: Project[];
+  loadingProjects: boolean;
   superuserNeedsToBeProjectMember?: boolean;
 }>;
 
-export default function NoProjectMessage({
+function NoProjectMessage({
   children,
   organization,
   projects,
@@ -127,3 +128,5 @@ const Content = styled(Flex)`
 const Actions = styled(ButtonBar)`
   width: fit-content;
 `;
+
+export default withProjects(NoProjectMessage);
