+++ static/app/views/performance/compare/index.tsx
@@ -4,9 +4,9 @@ import styled from '@emotion/styled';
 import * as Sentry from '@sentry/react';
 
 import NotFound from 'app/components/errors/notFound';
-import LightWeightNoProjectMessage from 'app/components/lightWeightNoProjectMessage';
 import LoadingError from 'app/components/loadingError';
 import LoadingIndicator from 'app/components/loadingIndicator';
+import NoProjectMessage from 'app/components/noProjectMessage';
 import SentryDocumentTitle from 'app/components/sentryDocumentTitle';
 import {t} from 'app/locale';
 import {PageContent} from 'app/styles/organization';
@@ -125,9 +125,9 @@ class TransactionComparisonPage extends React.PureComponent<Props> {
       >
         <React.Fragment>
           <StyledPageContent>
-            <LightWeightNoProjectMessage organization={organization}>
+            <NoProjectMessage organization={organization}>
               {this.renderComparison({baselineEventSlug, regressionEventSlug})}
-            </LightWeightNoProjectMessage>
+            </NoProjectMessage>
           </StyledPageContent>
         </React.Fragment>
       </SentryDocumentTitle>
