+++ static/app/views/releases/detail/index.tsx
@@ -6,8 +6,8 @@ import moment from 'moment';
 
 import Alert from 'app/components/alert';
 import AsyncComponent from 'app/components/asyncComponent';
-import LightWeightNoProjectMessage from 'app/components/lightWeightNoProjectMessage';
 import LoadingIndicator from 'app/components/loadingIndicator';
+import NoProjectMessage from 'app/components/noProjectMessage';
 import GlobalSelectionHeader from 'app/components/organizations/globalSelectionHeader';
 import {getParams} from 'app/components/organizations/globalSelectionHeader/getParams';
 import PickProjectToContinue from 'app/components/pickProjectToContinue';
@@ -196,7 +196,7 @@ class ReleasesDetail extends AsyncView<Props, State> {
     }
 
     return (
-      <LightWeightNoProjectMessage organization={organization}>
+      <NoProjectMessage organization={organization}>
         <StyledPageContent>
           <ReleaseHeader
             location={location}
@@ -223,7 +223,7 @@ class ReleasesDetail extends AsyncView<Props, State> {
             {this.props.children}
           </ReleaseContext.Provider>
         </StyledPageContent>
-      </LightWeightNoProjectMessage>
+      </NoProjectMessage>
     );
   }
 }
